pyproject.toml
[tool.poetry]
name = "fastapi-realworld-demo"
version = "0.1.0"
description = "A minimal FastAPI project."
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
alembic = "^1.16.2"
asyncpg = "^0.30.0"
email-validator = "^2.2.0"
fastapi = "^0.115.13"
pydantic-settings = "^2.9.1"
pyjwt = "^2.10.1"
python = "^3.12"
uvicorn = "^0.34.3"
bcrypt = "<4.1.0"
passlib = "^1.7.4"
types-passlib = "^1.7.7.20250602"
python-slugify = "^8.0.4"

[tool.poetry.group.dev.dependencies]
faker = "^37.4.0"
httpx = "^0.27.0"
mypy = "^1.0.0"
pytest = "<8.0.0"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.14.1"
ruff = "^0.12.0"
pytest-cov = "^6.2.1"
aiosqlite = "^0.21.0"
asgi-lifespan = "^2.1.0"
polyfactory = "^2.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
exclude = [".venv", "alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
